Validate create index requests' number of primary/replica shards
Fixes #7495
