+++ src/main/java/org/elasticsearch/action/admin/indices/create/CreateIndexRequest.java
@@ -107,6 +107,14 @@ public class CreateIndexRequest extends AcknowledgedRequest<CreateIndexRequest> 
         if (index == null) {
             validationException = addValidationError("index is missing", validationException);
         }
+        Integer number_of_primaries = settings.getAsInt(IndexMetaData.SETTING_NUMBER_OF_SHARDS, null);
+        Integer number_of_replicas = settings.getAsInt(IndexMetaData.SETTING_NUMBER_OF_REPLICAS, null);
+        if (number_of_primaries != null && number_of_primaries <= 0) {
+            validationException = addValidationError("index must have 1 or more primary shards", validationException);
+        }
+        if (number_of_replicas != null && number_of_replicas < 0) {
+            validationException = addValidationError("index must have 0 or more replica shards", validationException);
+        }
         return validationException;
     }
