+++ src/test/java/org/elasticsearch/action/admin/indices/create/CreateIndexTests.java
@@ -19,6 +19,7 @@
 
 package org.elasticsearch.action.admin.indices.create;
 
+import org.elasticsearch.action.ActionRequestValidationException;
 import org.elasticsearch.action.admin.cluster.state.ClusterStateResponse;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
@@ -30,6 +31,9 @@ import org.elasticsearch.test.ElasticsearchIntegrationTest.ClusterScope;
 import org.elasticsearch.test.ElasticsearchIntegrationTest.Scope;
 import org.junit.Test;
 
+import java.util.HashMap;
+
+import static org.elasticsearch.common.xcontent.XContentFactory.jsonBuilder;
 import static org.hamcrest.Matchers.*;
 import static org.hamcrest.core.IsNull.notNullValue;
 
@@ -51,7 +55,7 @@ public class CreateIndexTests extends ElasticsearchIntegrationTest{
         assertThat(index, notNullValue());
         assertThat(index.creationDate(), equalTo(4l));
     }
-    
+
     @Test
     public void testCreationDate_Generated() {
         long timeBeforeRequest = System.currentTimeMillis();
@@ -70,4 +74,70 @@ public class CreateIndexTests extends ElasticsearchIntegrationTest{
         assertThat(index.creationDate(), allOf(lessThanOrEqualTo(timeAfterRequest), greaterThanOrEqualTo(timeBeforeRequest)));
     }
 
+    @Test
+    public void testDoubleAddMapping() throws Exception {
+        try {
+            prepareCreate("test")
+                    .addMapping("type1", "date", "type=date")
+                    .addMapping("type1", "num", "type=integer");
+            fail("did not hit expected exception");
+        } catch (IllegalStateException ise) {
+            // expected
+        }
+        try {
+            prepareCreate("test")
+                    .addMapping("type1", new HashMap<String,Object>())
+                    .addMapping("type1", new HashMap<String,Object>());
+            fail("did not hit expected exception");
+        } catch (IllegalStateException ise) {
+            // expected
+        }
+        try {
+            prepareCreate("test")
+                    .addMapping("type1", jsonBuilder())
+                    .addMapping("type1", jsonBuilder());
+            fail("did not hit expected exception");
+        } catch (IllegalStateException ise) {
+            // expected
+        }
+    }
+
+    @Test
+    public void testInvalidShardCountSettings() throws Exception {
+        try {
+            prepareCreate("test").setSettings(ImmutableSettings.builder()
+                    .put(IndexMetaData.SETTING_NUMBER_OF_SHARDS, randomIntBetween(-10, 0))
+                    .build())
+            .get();
+            fail("should have thrown an exception about the primary shard count");
+        } catch (ActionRequestValidationException e) {
+            assertThat("message contains error about shard count: " + e.getMessage(),
+                    e.getMessage().contains("index must have 1 or more primary shards"), equalTo(true));
+        }
+
+        try {
+            prepareCreate("test").setSettings(ImmutableSettings.builder()
+                    .put(IndexMetaData.SETTING_NUMBER_OF_REPLICAS, randomIntBetween(-10, -1))
+                    .build())
+                    .get();
+            fail("should have thrown an exception about the replica shard count");
+        } catch (ActionRequestValidationException e) {
+            assertThat("message contains error about shard count: " + e.getMessage(),
+                    e.getMessage().contains("index must have 0 or more replica shards"), equalTo(true));
+        }
+
+        try {
+            prepareCreate("test").setSettings(ImmutableSettings.builder()
+                    .put(IndexMetaData.SETTING_NUMBER_OF_SHARDS, randomIntBetween(-10, 0))
+                    .put(IndexMetaData.SETTING_NUMBER_OF_REPLICAS, randomIntBetween(-10, -1))
+                    .build())
+                    .get();
+            fail("should have thrown an exception about the shard count");
+        } catch (ActionRequestValidationException e) {
+            assertThat("message contains error about shard count: " + e.getMessage(),
+                    e.getMessage().contains("index must have 1 or more primary shards"), equalTo(true));
+            assertThat("message contains error about shard count: " + e.getMessage(),
+                    e.getMessage().contains("index must have 0 or more replica shards"), equalTo(true));
+        }
+    }
 }
